# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FanVerse - Football Fan Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discussions:
    get:
      operationId: listDiscussions
      tags: [discussions]
      summary: List community discussions
      parameters:
        - name: nationCode
          in: query
          schema:
            type: string
        - name: matchId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of discussions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Discussion"
    post:
      operationId: createDiscussion
      tags: [discussions]
      summary: Create a new discussion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscussionInput"
      responses:
        "201":
          description: Discussion created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
        "401":
          description: Unauthorized
  /discussions/{id}:
    get:
      operationId: getDiscussion
      tags: [discussions]
      summary: Get discussion with comments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discussion with comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscussionDetail"
        "404":
          description: Not found
  /discussions/{id}/comments:
    post:
      operationId: addComment
      tags: [discussions]
      summary: Add a comment to a discussion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
  /discussions/{id}/upvote:
    post:
      operationId: upvoteDiscussion
      tags: [discussions]
      summary: Upvote a discussion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Upvoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Discussion"
        "401":
          description: Unauthorized
components:
  schemas:
...
    Discussion:
      type: object
      properties:
        id:
          type: integer
        nationCode:
          type: ["string", "null"]
        matchId:
          type: ["integer", "null"]
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        userNationCode:
          type: ["string", "null"]
        reputationTier:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [analysis, hype, prediction, reaction, history, general]
        upvotes:
          type: integer
        commentCount:
          type: integer
        hasUserUpvoted:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, username, reputationTier, title, content, category, upvotes, commentCount, hasUserUpvoted, createdAt, updatedAt]
    DiscussionInput:
      type: object
      properties:
        nationCode:
          type: string
        matchId:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
          enum: [analysis, hype, prediction, reaction, history, general]
      required: [title, content, category]
    DiscussionDetail:
      type: object
      properties:
        id:
          type: integer
        nationCode:
          type: ["string", "null"]
        matchId:
          type: ["integer", "null"]
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        userNationCode:
          type: ["string", "null"]
        reputationTier:
          type: string
        title:
          type: string
        content:
          type: string
        category:
          type: string
        upvotes:
          type: integer
        commentCount:
          type: integer
        hasUserUpvoted:
          type: boolean
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, userId, username, reputationTier, title, content, category, upvotes, commentCount, hasUserUpvoted, comments, createdAt, updatedAt]
    Comment:
      type: object
      properties:
        id:
          type: integer
        discussionId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        userNationCode:
          type: ["string", "null"]
        reputationTier:
          type: string
        content:
          type: string
        upvotes:
          type: integer
        createdAt:
          type: string
      required: [id, discussionId, userId, username, reputationTier, content, upvotes, createdAt]
    CommentInput:
      type: object
      properties:
        content:
          type: string
      required: [content]