# lib/api-zod/src/generated/types/discussionDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanVerse - Football Fan Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Comment } from './comment';

export interface DiscussionDetail {
  id: number;
  /** @nullable */
  nationCode?: string | null;
  /** @nullable */
  matchId?: number | null;
  userId: number;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  userNationCode?: string | null;
  reputationTier: string;
  title: string;
  content: string;
  category: string;
  upvotes: number;
  commentCount: number;
  hasUserUpvoted: boolean;
  comments: Comment[];
  createdAt: string;
  updatedAt: string;
}