# lib/api-zod/src/generated/types/discussionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FanVerse - Football Fan Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DiscussionInputCategory } from './discussionInputCategory';

export interface DiscussionInput {
  nationCode?: string;
  matchId?: number;
  title: string;
  content: string;
  category: DiscussionInputCategory;
}